Move Results into results.py Beautiful output results in test case
